+++ test/rules/test_md022.py
@@ -1925,6 +1925,416 @@ def test_md022_fenced_block_before_header():
     )
 
 
+@pytest.mark.rules
+def test_md022_issue_1268():
+    """
+    Test to make sure this rule
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    source_file_contents = """# My header
+
+<!-- pyml disable-num-lines 3 md011,md013-->
+
+Some long lines
+"""
+    with create_temporary_configuration_file(
+        source_file_contents, file_name_suffix=".md"
+    ) as source_path:
+        supplied_arguments = [
+            "scan",
+            source_path,
+        ]
+
+        expected_return_code = 0
+        expected_output = ""
+        expected_error = ""
+
+        # Act
+        execute_results = scanner.invoke_main(arguments=supplied_arguments)
+
+        # Assert
+        execute_results.assert_results(
+            expected_output, expected_error, expected_return_code
+        )
+
+
+@pytest.mark.rules
+def test_md022_good_with_pragmas_blanks_after_around_and_before():
+    """
+    Test to make sure this rule works properly with pragmas.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    source_file_contents = """# Heading 1
+
+<!--pyml disable-num-lines 1 md012-->
+
+Some text
+
+Some more text
+
+<!--pyml disable-num-lines 1 md012-->
+
+## Heading 2
+
+<!--pyml disable-num-lines 5 md012-->
+
+## Heading 3
+"""
+    with create_temporary_configuration_file(
+        source_file_contents, file_name_suffix=".md"
+    ) as source_path:
+        supplied_configuration = {
+            "plugins": {"md022": {"lines_below": 1, "lines_above": 1}}
+        }
+        with create_temporary_configuration_file(
+            supplied_configuration
+        ) as configuration_file:
+            supplied_arguments = [
+                "-c",
+                configuration_file,
+                "scan",
+                source_path,
+            ]
+
+            expected_return_code = 0
+            expected_output = ""
+            expected_error = ""
+
+            # Act
+            execute_results = scanner.invoke_main(arguments=supplied_arguments)
+
+            # Assert
+            execute_results.assert_results(
+                expected_output, expected_error, expected_return_code
+            )
+
+
+@pytest.mark.rules
+def test_md022_bad_with_pragmas_blanks_after_around_and_before():
+    """
+    Test to make sure this rule works properly with pragmas.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    source_file_contents = """# Heading 1
+
+<!--pyml disable-num-lines 1 md012-->
+
+Some text
+
+Some more text
+
+<!--pyml disable-num-lines 1 md012-->
+
+## Heading 2
+
+<!--pyml disable-num-lines 5 md012-->
+
+## Heading 3
+"""
+    with create_temporary_configuration_file(
+        source_file_contents, file_name_suffix=".md"
+    ) as source_path:
+        supplied_configuration = {
+            "plugins": {"md022": {"lines_below": 2, "lines_above": 2}}
+        }
+        with create_temporary_configuration_file(
+            supplied_configuration
+        ) as configuration_file:
+            supplied_arguments = [
+                "-c",
+                configuration_file,
+                "scan",
+                source_path,
+            ]
+
+            expected_return_code = 1
+            expected_output = f"""{source_path}:1:1: MD022: Headings should be surrounded by blank lines. [Expected: 2; Actual: 1; Below] (blanks-around-headings,blanks-around-headers)
+{source_path}:11:1: MD022: Headings should be surrounded by blank lines. [Expected: 2; Actual: 1; Above] (blanks-around-headings,blanks-around-headers)
+{source_path}:11:1: MD022: Headings should be surrounded by blank lines. [Expected: 2; Actual: 1; Below] (blanks-around-headings,blanks-around-headers)
+{source_path}:15:1: MD022: Headings should be surrounded by blank lines. [Expected: 2; Actual: 1; Above] (blanks-around-headings,blanks-around-headers)
+{source_path}:15:1: MD022: Headings should be surrounded by blank lines. [Expected: 2; Actual: 1; Below] (blanks-around-headings,blanks-around-headers)"""
+            expected_error = ""
+
+            # Act
+            execute_results = scanner.invoke_main(arguments=supplied_arguments)
+
+            # Assert
+            execute_results.assert_results(
+                expected_output, expected_error, expected_return_code
+            )
+
+
+@pytest.mark.rules
+def test_md022_good_with_pragmas_blanks_before():
+    """
+    Test to make sure this rule works properly with pragmas.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    source_file_contents = """# Heading 1
+
+<!--pyml disable-num-lines 1 md012-->
+
+Some text
+"""
+    with create_temporary_configuration_file(
+        source_file_contents, file_name_suffix=".md"
+    ) as source_path:
+        supplied_configuration = {
+            "plugins": {"md022": {"lines_below": 1, "lines_above": 1}}
+        }
+        with create_temporary_configuration_file(
+            supplied_configuration
+        ) as configuration_file:
+            supplied_arguments = [
+                "-c",
+                configuration_file,
+                "scan",
+                source_path,
+            ]
+
+            expected_return_code = 0
+            expected_output = ""
+            expected_error = ""
+
+            # Act
+            execute_results = scanner.invoke_main(arguments=supplied_arguments)
+
+            # Assert
+            execute_results.assert_results(
+                expected_output, expected_error, expected_return_code
+            )
+
+
+@pytest.mark.rules
+def test_md022_bad_with_pragmas_blanks_before():
+    """
+    Test to make sure this rule works properly with pragmas.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    source_file_contents = """# Heading 1
+
+<!--pyml disable-num-lines 1 md012-->
+
+Some text
+"""
+    with create_temporary_configuration_file(
+        source_file_contents, file_name_suffix=".md"
+    ) as source_path:
+        supplied_configuration = {
+            "plugins": {"md022": {"lines_below": 2, "lines_above": 2}}
+        }
+        with create_temporary_configuration_file(
+            supplied_configuration
+        ) as configuration_file:
+            supplied_arguments = [
+                "-c",
+                configuration_file,
+                "scan",
+                source_path,
+            ]
+
+            expected_return_code = 1
+            expected_output = f"""{source_path}:1:1: MD022: Headings should be surrounded by blank lines. [Expected: 2; Actual: 1; Below] (blanks-around-headings,blanks-around-headers)"""
+            expected_error = ""
+
+            # Act
+            execute_results = scanner.invoke_main(arguments=supplied_arguments)
+
+            # Assert
+            execute_results.assert_results(
+                expected_output, expected_error, expected_return_code
+            )
+
+
+@pytest.mark.rules
+def test_md022_good_with_pragmas_blanks_around():
+    """
+    Test to make sure this rule works properly with pragmas.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    source_file_contents = """Some text
+
+<!--pyml disable-num-lines 1 md012-->
+
+## Heading 2
+
+<!--pyml disable-num-lines 5 md012-->
+
+Some more text
+"""
+    with create_temporary_configuration_file(
+        source_file_contents, file_name_suffix=".md"
+    ) as source_path:
+        supplied_configuration = {
+            "plugins": {"md022": {"lines_below": 1, "lines_above": 1}}
+        }
+        with create_temporary_configuration_file(
+            supplied_configuration
+        ) as configuration_file:
+            supplied_arguments = [
+                "-c",
+                configuration_file,
+                "scan",
+                source_path,
+            ]
+
+            expected_return_code = 0
+            expected_output = ""
+            expected_error = ""
+
+            # Act
+            execute_results = scanner.invoke_main(arguments=supplied_arguments)
+
+            # Assert
+            execute_results.assert_results(
+                expected_output, expected_error, expected_return_code
+            )
+
+
+@pytest.mark.rules
+def test_md022_bad_with_pragmas_blanks_around():
+    """
+    Test to make sure this rule works properly with pragmas.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    source_file_contents = """Some text
+
+<!--pyml disable-num-lines 1 md012-->
+
+## Heading 2
+
+<!--pyml disable-num-lines 5 md012-->
+
+Some more text
+"""
+    with create_temporary_configuration_file(
+        source_file_contents, file_name_suffix=".md"
+    ) as source_path:
+        supplied_configuration = {
+            "plugins": {"md022": {"lines_below": 2, "lines_above": 2}}
+        }
+        with create_temporary_configuration_file(
+            supplied_configuration
+        ) as configuration_file:
+            supplied_arguments = [
+                "-c",
+                configuration_file,
+                "scan",
+                source_path,
+            ]
+
+            expected_return_code = 1
+            expected_output = f"""{source_path}:5:1: MD022: Headings should be surrounded by blank lines. [Expected: 2; Actual: 1; Above] (blanks-around-headings,blanks-around-headers)
+{source_path}:5:1: MD022: Headings should be surrounded by blank lines. [Expected: 2; Actual: 1; Below] (blanks-around-headings,blanks-around-headers)"""
+            expected_error = ""
+
+            # Act
+            execute_results = scanner.invoke_main(arguments=supplied_arguments)
+
+            # Assert
+            execute_results.assert_results(
+                expected_output, expected_error, expected_return_code
+            )
+
+
+@pytest.mark.rules
+def test_md022_good_with_pragmas_blanks_after():
+    """
+    Test to make sure this rule works properly with pragmas.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    source_file_contents = """<!--pyml disable-num-lines 5 md012-->
+
+## Heading 3
+"""
+    with create_temporary_configuration_file(
+        source_file_contents, file_name_suffix=".md"
+    ) as source_path:
+        supplied_configuration = {
+            "plugins": {"md022": {"lines_below": 1, "lines_above": 1}}
+        }
+        with create_temporary_configuration_file(
+            supplied_configuration
+        ) as configuration_file:
+            supplied_arguments = [
+                "-c",
+                configuration_file,
+                "scan",
+                source_path,
+            ]
+
+            expected_return_code = 0
+            expected_output = ""
+            expected_error = ""
+
+            # Act
+            execute_results = scanner.invoke_main(arguments=supplied_arguments)
+
+            # Assert
+            execute_results.assert_results(
+                expected_output, expected_error, expected_return_code
+            )
+
+
+@pytest.mark.rules
+def test_md022_bad_with_pragmas_blanks_after():
+    """
+    Test to make sure this rule works properly with pragmas.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    source_file_contents = """
+<!--pyml disable-num-lines 5 md012-->
+
+## Heading 3
+"""
+    with create_temporary_configuration_file(
+        source_file_contents, file_name_suffix=".md"
+    ) as source_path:
+        supplied_configuration = {
+            "plugins": {"md022": {"lines_below": 2, "lines_above": 2}}
+        }
+        with create_temporary_configuration_file(
+            supplied_configuration
+        ) as configuration_file:
+            supplied_arguments = [
+                "-c",
+                configuration_file,
+                "scan",
+                source_path,
+            ]
+
+            expected_return_code = 1
+            expected_output = f"""{source_path}:4:1: MD022: Headings should be surrounded by blank lines. [Expected: 2; Actual: 1; Above] (blanks-around-headings,blanks-around-headers)
+{source_path}:4:1: MD022: Headings should be surrounded by blank lines. [Expected: 2; Actual: 1; Below] (blanks-around-headings,blanks-around-headers)"""
+            expected_error = ""
+
+            # Act
+            execute_results = scanner.invoke_main(arguments=supplied_arguments)
+
+            # Assert
+            execute_results.assert_results(
+                expected_output, expected_error, expected_return_code
+            )
+
+
 def test_md022_query_config():
     config_test = pluginQueryConfigTest(
         "md022",